Move TextLines type to its own file
